move launcher to eastier lib
